# Pertemuan12/Tugas/Praktikum_714230047/docs/swagger.yaml
basePath: /
definitions:
...
  model.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  model.LoginResponse:
    properties:
      message:
        example: Login success
        type: string
      token:
        example: v4.public.xxxxxx
        type: string
    type: object
...
  model.UserLogin:
    properties:
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:8088
info:
  contact:
    email: indra@ulbi.ac.id
    name: API Support
    url: https://github.com/indrariksa
  description: This is a sample swagger for Fiber
  title: TES SWAGGER PEMROGRAMAN III
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Melakukan proses login dan mengembalikan token PASETO jika username
        dan password valid.
      parameters:
      - description: Login credentials (username dan password)
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login success
          schema:
            $ref: '#/definitions/model.LoginResponse'
        "400":
          description: Invalid body
        "401":
          description: Username not found or Wrong password
        "500":
          description: Failed to generate token
      summary: Login User
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      description: Melakukan proses registrasi user baru dengan username, password,
        dan role.
      parameters:
      - description: Register credentials (username, password, dan role)
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/model.UserLogin'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input atau Username, password, and role are required
        "409":
          description: User already exists
        "500":
          description: Failed to hash password
      summary: Register User
      tags:
      - Auth
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"